requirements.txt
aiohttp==3.9.2
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
click==8.1.8
dataclasses-json==0.6.1
fastapi==0.110.1
frozenlist==1.4.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.4
jsonpatch==1.33
jsonpointer==2.4
marshmallow==3.20.1
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.26.0
packaging==23.2
python-dotenv==1.0.0
PyYAML==6.0.2
requests==2.32.3
sniffio==1.3.0
sse-starlette==1.6.5
starlette==0.37.2
tenacity==8.2.3
tqdm==4.67.1
typing-inspect==0.9.0
typing_extensions==4.12.0
urllib3==1.26.15
uvicorn==0.23.2
yarl==1.9.2
cachetools>=4.2.2
pyjwt~=2.8.0
pandas==2.2.2
python-multipart==0.0.9
requests-toolbelt==1.0.0
accelerate>=0.26.0
markitdown[pdf,docx,xlsx,xls,pptx]
onnxocr==2025.5
ollama==0.5.2
dashscope==1.24.1